third_party/blink/web_tests/fast/filesystem/read-directory-many-expected.txt

DirectoryReader.readEntries() test with many entries.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Successfully obtained TEMPORARY FileSystem:file__0:Temporary
PASS numFiles is numFilesExpected
PASS numDirectories is numDirectoriesExpected
PASS successfullyParsed is true

TEST COMPLETE
